mini-NoteDemo/djagno后台/mynote/requirements.txt
appnope==0.1.0
argh==0.26.2
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==17.3.0
autobahn==18.3.1
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==2.1.3
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
click==6.7
cloudinary==1.11.0
configparser==3.5.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
cycler==0.10.0
daphne==1.4.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-redis==4.10.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
html5lib==1.0.1
httplib2==0.12.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.2
mock==2.0.0
msgpack-python==0.5.6
Naked==0.1.31
numpy==1.16.1
odfpy==1.4.0
openpyxl==2.6.0
pbr==3.1.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.18
pycryptodome==3.7.3
PyHamcrest==1.9.0
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
reversion==0.2
shellescape==3.4.1
six==1.11.0
tablib==0.12.1
toml==0.10.0
Twisted==18.9.0
txaio==2.9.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.4.3
